(move) importmap implementation to the plugin
