+++ core/tests/effects.test.ts
@@ -2,7 +2,8 @@ import { assertEquals } from "@std/assert/equals";
 import { assertGreaterOrEqual } from "@std/assert/greater-or-equal";
 import { assertLessOrEqual } from "@std/assert/less-or-equal";
 import { createEffect, handlerFor, runWith } from "../src/effects/effects.ts";
-import { Handler, id } from "../src/utils/algebraic-structures.ts";
+import { id } from "../src/utils/algebraic-structures.ts";
+import { Handler } from "../src/effects/handlers.ts";
 
 /**
  * Effect definitions
